Major update
Change touch-cursor according to google-dev-tools. Add a google map.
Change the style of phone and tablet according to iPad and iPhone5
design. To be precise, add an iPhone5 button, speaker and camera,
change the size and colors. Change the background color. Add a
fit-to-screen function if the device is larger than the screen.
Refactor the device mark-up from body to a div with a class ‘device’.
Change the mark-up and style of resume. To be precise, give 100% size
to the resume page. Change some svgs. Hide the footer svgs, show them
only on hover. Change document.querySelector() to s() and
document.querySelectorAll() to S(). Change the resume navigation.
Disable drag-scroll on a map. Change the animation time of bounceInLeft
to 1s. Hide the resume-navigation-arrows for devices with width less
than 768px. Change some colors.